requirements.txt
discord.py>=2.3.0
sqlalchemy>=2.0.0
aiomysql>=0.2.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pytz>=2023.3
Pillow>=10.0.0
pytesseract>=0.3.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
aiohttp>=3.9.0